backend/telegram-backend/requirements.txt
telethon
python-dotenv
fastapi
uvicorn
